Make wave particle shader output to LodData
The output is broken, and will break the gestner waves as well. But it's
a good first step for experimenting.

Whilst I expect that WaveParticles will eventually move to not using
compute shaders (due to ROPs being needed to blend particle outputs,
maybe), in the short term they will as that is what my old
implementation uses. I will stick to that for now, as it will force a
proper think about how to support compute shaders and traditional
shaders through a LodDataInput system, which will be useful for other
things.

Such as ironically, supporting compute ShapeGerstnerBatched and
traditionally-rendered wave particles.