Change underwater post-process approach
The goal is to now have two quads, each with a different material
attached to them. The goal of the first quad is to stencil-out
everything up to the horizon line. The ocean is then responsible for
stenciling in-and-out areas as appropriate (depending on ocean face that
is being rendered) before the post process effect is applied to the
stencilled area.